+++ frontend/src/components/LoginButton/LoginButton.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useAuth0 } from "@auth0/auth0-react";
-import styles from "../styles/LoginButton.module.css";
+import styles from "./LoginButton.module.css";
 
 const LoginButton = () => {
   const { loginWithRedirect } = useAuth0();
